ai-service/requirements.txt
# Minimal dependencies - add more as needed
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
ruff==0.1.6

# Add your AI/ML dependencies here as needed:
# openai==1.3.0
# langchain==0.1.0
# chromadb==0.4.18
# tiktoken==0.5.1
# sentence-transformers==2.2.2
# numpy==1.24.3
# pandas==2.1.3
# pypdf==3.17.1
# python-multipart==0.0.6